pyproject.toml
[project]
name = "galadriel"
version = "0.1.0"
description = "GALADRIEL - Generative Attention-based Learning And Delay Recognition In Event Logs"
authors = [
    { name = "Simone Garau", email = "20005068@studenti.uniupo.it" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pm4py>=2.7.0",
    "scikit-learn>=1.3.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "tokenizers>=0.13.0",
    "datasets>=2.14.0",
    "requests>=2.31.0",
    "httpx>=0.24.0",
    "google-generativeai>=0.3.0",
    "captum>=0.6.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "jupyter>=1.0.0",
    "ipywidgets>=8.0.0",
    "aiohttp>=3.8.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "pyyaml>=6.0",
    "ollama",
    # Optimization & Experiment Tracking
    "optuna>=3.4.0",
    "optuna-dashboard>=0.14.0",
    "mlflow>=2.8.0",
    "accelerate>=1.9.0",
    "bertviz>=1.4.1",
    "tiktoken>=0.11.0",
    "pystache>=0.6.8",
    "langchain>=0.3.27",
    "langchain-community>=0.3.21",
    "langchain-chroma>=0.2.2",
    "langchain-ollama>=0.3.7",
    "umap-learn>=0.5.9.post2",
]

 [project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
]

agent = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.30",
    "langchain-chroma>=0.1.0",
    "chromadb>=0.4.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 79
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "pytest>=8.4.1",
]
